Phase 2: the Pay page IS the bill rolodex
Bills were two taps deep — a tile that opened a modal that listed them. Now they're
the page: BillRolodex renders the full list inline with search, add, per-bill quick
actions and tap-to-detail.

- New components/app-ui/BillRolodex.tsx; the add-a-biller form (provider autofill,
  type grid, all fields) moves inline rather than living in a modal.
- The per-bill quick action adapts to verification: verified members pay from Cash,
  everyone else is sent to the biller's own site (disabled when there's no link).
- BillPortalsModal and PayContext.openPortals are deleted — one surface, so there's
  no second bill list to keep in sync.
- The "Pay a bill" tile is gone; the remaining tiles are now "Quick actions".
